+++ services/mailer/message.go
@@ -41,7 +41,7 @@ func (s *Server) ListThreadMessages(
 		userInfo,
 		req.GetThreadId(),
 		req.GetEmailId(),
-		maileraccess.AccessLevel_ACCESS_LEVEL_READ,
+		int32(maileraccess.AccessLevel_ACCESS_LEVEL_READ),
 	); err != nil {
 		return nil, err
 	}
@@ -165,7 +165,7 @@ func (s *Server) PostMessage(
 		userInfo,
 		req.GetMessage().GetThreadId(),
 		req.GetMessage().GetSenderId(),
-		maileraccess.AccessLevel_ACCESS_LEVEL_WRITE,
+		int32(maileraccess.AccessLevel_ACCESS_LEVEL_WRITE),
 	); err != nil {
 		return nil, err
 	}
